Add docs, make Category a pure model class
